requirements.txt
fastapi
uvicorn
ctransformers
sentence-transformers
requests
streamlit
langchain
langchain-community
langsmith
literalai
faiss-cpu
streamlit
python-dotenv
mistralai 
tiktoken 
pypdf 
pytesseract
opencv-python
pdf2image
pillow
pdfplumber
python-docx
python-multipart
numpy
groq